refactor: Extract Film row mapper to a separate class
